requirements.txt
# HTTP client
requests>=2.32.3

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Configuration
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.0
asyncio>=3.4.3

# Scraping & Automation
playwright>=1.42.0
playwright-stealth>=1.0.6
crawl4ai>=0.2.0
aiohttp>=3.9.0
scrapling>=0.2.9
curl_cffi>=0.7.0
w3lib>=2.1.0

# API-based scrapers
apify-client>=1.7.0
firecrawl-py>=1.0.0

# Google Sheets API
gspread>=6.0.0
google-auth>=2.29.0

# --- NEW: Hybrid Workflow ---

# Scrapy spider framework
Scrapy>=2.11.0
scrapy-playwright>=0.0.34

# MongoDB async driver
motor>=3.4.0
pymongo>=4.7.0

# NLP / Sentiment (Tamil-aware)
transformers>=4.40.0
sentencepiece>=0.2.0
torch>=2.2.0
# CPU-only torch is sufficient; GPU optional
# If you have CUDA: pip install torch --index-url https://download.pytorch.org/whl/cu121